fix: Resolve build errors from GUID refactoring
This commit fixes a series of build errors that arose after refactoring the application to support GUIDs for private chats.

The following issues were addressed:
- **CS1513 & CS0161 in `ChatHub.cs`:** Fixed a syntax error (mismatched brace) and a missing return path in the `DetermineIfShouldQueue` method.
- **CS0029 in `ChatHub.cs`:** Corrected a type mismatch by refactoring the `ParseGroupKey` method to return a `long` instead of `int`, ensuring type consistency.
- **CS0535 in `MessageService.cs`:** Updated the `EditChannelMessageAsync` method signature to match the `IMessageService` interface.
- **CS0019 & CS0029 in `MessageService.cs`:** Fixed various type mismatch errors by consistently using the parsed numeric ID (`long`) instead of the raw string ID where a number was expected.
- **CS1503 in `BroadcastService.cs` & `ProcessMessageJob.cs`:**
    - Modified `QueuedMessageDto` to accept a `string` GroupId.
    - Updated `ProcessMessageJob.cs` to correctly parse the string GroupId to a long for internal operations.
    - Corrected a method call in `BroadcastService.cs` to pass a string ID to the message service.

+++ Messenger.API/Hubs/ChatHub.cs
@@ -424,16 +424,16 @@ namespace Messenger.API.Hubs
             return allUsers.Select(u => new { UserId = u.UserId, UserName = u.NameFamily, ProfilePic = u.ProfilePicName, IsOnline = onlineSet.Contains(u.UserId) }).Cast<object>().ToList();
         }
 
-        private (int groupId, string groupType) ParseGroupKey(string groupKey)
+        private (long groupId, string groupType) ParseGroupKey(string groupKey)
         {
             if (groupKey.StartsWith(ConstChat.ClassGroup))
             {
-                int.TryParse(groupKey.Substring(ConstChat.ClassGroup.Length), out int id);
+                long.TryParse(groupKey.Substring(ConstChat.ClassGroup.Length), out long id);
                 return (id, ConstChat.ClassGroupType);
             }
             if (groupKey.StartsWith(ConstChat.ChannelGroup))
             {
-                int.TryParse(groupKey.Substring(ConstChat.ChannelGroup.Length), out int id);
+                long.TryParse(groupKey.Substring(ConstChat.ChannelGroup.Length), out long id);
                 return (id, ConstChat.ChannelGroupType);
             }
             return (0, string.Empty);
